Fix issue with Replace rules that are prefixes
If the replacement of one Replace rule is a prefix
of the target of another Replace rule, then there
are unexpected results. Consider

Replace,c̣,čh
Replace,c,č

The target of the second rule, "c" is a prefix of the
replacement in the first rule "čh" since we use the
Canonical Decomposition of all Unicode strings.
This is likely to occur since there are numerous
combining diacritic marks in use.

In order to address this, add a suffix to target
patterns that prevents them from matching a Unicode
sequence that would end in the middle of a sequence
of combining diacritic characters.